Move survey related hooks from smartSearch to surveys
