+++ src/features/views/components/ViewColumnDialog/SurveyResponsePluralConfig.tsx
@@ -11,7 +11,7 @@ import {
 
 import messageIds from 'features/views/l10n/messageIds';
 import { useNumericRouteParams } from 'core/hooks';
-import useSurveysWithElements from 'features/smartSearch/hooks/useSurveysWithElements';
+import useSurveysWithElements from 'features/surveys/hooks/useSurveysWithElements';
 import ZUIFuture from 'zui/ZUIFuture';
 
 interface SurveyResponsePluralConfigProps {
